refactor tensorops into separate files
