+++ core/src/main/scala/dimwit/tensor/tensorops/Tensor0Ops.scala
@@ -1,0 +1,61 @@
+package dimwit.tensor.tensorops
+
+import dimwit.tensor.Tensor0
+import dimwit.tensor.DType.*
+import dimwit.tensor.TensorOps
+import dimwit.tensor.TensorOps.IsNumber
+import dimwit.tensor.Labels
+import dimwit.tensor.tensorops.ElementWiseOps.add
+import dimwit.tensor.Tensor
+import dimwit.tensor.tensorops.TensorOpsUtil.Broadcast
+import dimwit.tensor.tensorops.ElementWiseOps.subtract
+import dimwit.tensor.tensorops.ElementWiseOps.multiply
+import dimwit.tensor.tensorops.ElementWiseOps.divide
+import dimwit.tensor.TensorOps.IsFloating
+
+object Tensor0Ops:
+
+  private inline def checkTracer[V, R](scalar: Tensor0[V]): Unit =
+    require(
+      !scalar.isTracer,
+      """
+        | Cannot convert a JAX Tracer to a scalar value. Tensor0 is part of a JAX computation graph (e.g., inside vmap or a jitted function).
+        | Common mistakes leading to this error:
+        |   - calling .slice(t0.item) rather than .slice(t0); breaking the computation graph unintentionally.
+        |""".stripMargin
+    )
+
+  extension (scalar: Tensor0[Bool])
+    def item: Boolean =
+      checkTracer(scalar)
+      scalar.jaxValue.item().as[Boolean]
+
+  extension (scalar: Tensor0[Int8])
+    def item: Byte =
+      checkTracer(scalar)
+      scalar.jaxValue.item().as[Byte]
+
+  extension (scalar: Tensor0[Int16])
+    def item: Short =
+      checkTracer(scalar)
+      scalar.jaxValue.item().as[Int].toShort
+
+  extension (scalar: Tensor0[Int32])
+    def item: Int =
+      checkTracer(scalar)
+      scalar.jaxValue.item().as[Int]
+
+  extension (scalar: Tensor0[Int64])
+    def item: Long =
+      checkTracer(scalar)
+      scalar.jaxValue.item().as[Long]
+
+  extension (scalar: Tensor0[Float32])
+    def item: Float =
+      checkTracer(scalar)
+      scalar.jaxValue.item().as[Float]
+
+  extension (scalar: Tensor0[Float64])
+    def item: Double =
+      checkTracer(scalar)
+      scalar.jaxValue.item().as[Double]
